refactor: change MajorType
