+++ src/main/java/com/gist/graduation/requirment/domain/GraduationRequirementStatus.java
@@ -4,6 +4,7 @@ import com.gist.graduation.requirment.domain.etc.EtcMandatory;
 import com.gist.graduation.requirment.domain.humanities.Humanities;
 import com.gist.graduation.requirment.domain.language.LanguageBasic;
 import com.gist.graduation.requirment.domain.major.Major;
+import com.gist.graduation.requirment.domain.major.MajorType;
 import com.gist.graduation.requirment.domain.other.OtherUncheckedClass;
 import com.gist.graduation.requirment.domain.science.ScienceBasic;
 import com.gist.graduation.user.taken_course.TakenCourse;
